reorganize into modules to add alternate implementations easier
